Move opening of Scan BP passport screen to Mobius
