+++ app/src/main/java/org/simple/clinic/home/patients/PatientsEffectHandler.kt
@@ -44,6 +44,7 @@ class PatientsEffectHandler @AssistedInject constructor(
         .addAction(ShowUserWasApproved::class.java, uiActions::showUserStatusAsApproved, schedulers.ui())
         .addAction(ShowUserPendingSmsVerification::class.java, uiActions::showUserStatusAsPendingVerification, schedulers.ui())
         .addAction(HideUserAccountStatus::class.java, uiActions::hideUserAccountStatus, schedulers.ui())
+        .addAction(OpenScanBpPassportScreen::class.java, uiActions::openScanSimpleIdCardScreen, schedulers.ui())
         .build()
   }
