+++ app/src/main/java/org/simple/clinic/home/patients/PatientsScreenEvents.kt
@@ -1,20 +1,6 @@
 package org.simple.clinic.home.patients
 
-import android.Manifest
-import org.simple.clinic.util.None
-import org.simple.clinic.util.Optional
-import org.simple.clinic.util.RequiresPermission
-import org.simple.clinic.platform.util.RuntimePermissionResult
 import org.simple.clinic.widgets.UiEvent
-
-data class ScanCardIdButtonClicked(
-    override var permission: Optional<RuntimePermissionResult> = None,
-    override val permissionString: String = Manifest.permission.CAMERA,
-    override val permissionRequestCode: Int = 1
-) : UiEvent, RequiresPermission {
-
-  override val analyticsName: String = "Patients:Scan Simple Card Clicked:"
-}
 
 object SimpleVideoClicked : UiEvent {
   override val analyticsName = "Patients:Simple Video Clicked"
