+++ app/src/main/java/org/simple/clinic/home/patients/PatientsUi.kt
@@ -1,7 +1,6 @@
 package org.simple.clinic.home.patients
 
 interface PatientsUi: PatientsUiActions {
-  fun openScanSimpleIdCardScreen()
   fun hideSyncIndicator()
   fun showSyncIndicator()
   fun showAppUpdateDialog()
